Removed LunarGVulkan.csproj, and fixed build issue in the unfinished tutorial (there still some good stuff in there).

+++ VulkanTutorial/SwapChain.cs
@@ -33,7 +33,7 @@ namespace VulkanTutorial
 		public void Initialize(IWindow window) {
 			CreateSurface(window);
 
-			NodeIndex = PhysicalDevice.GetQueueFamilyPropertiesExt()
+			NodeIndex = (int)PhysicalDevice.GetQueueFamilyPropertiesExt()
 			                          .First(q => q.SupportsGraphics && q.Supports(Surface))
 			                          .QueueIndex;
